refactor(core): split frontmatter enum schema module
